+++ hc/api/management/commands/smtpd.py
@@ -12,45 +12,49 @@ RE_UUID = re.compile(
 )
 
 
+def _process_message(remote_addr, mailfrom, mailto, data):
+    to_parts = mailto.split("@")
+    code = to_parts[0]
+
+    try:
+        data = data.decode()
+    except UnicodeError:
+        data = "[binary data]"
+
+    if not RE_UUID.match(code):
+        return f"Not an UUID: {code}"
+
+    try:
+        check = Check.objects.get(code=code)
+    except Check.DoesNotExist:
+        return f"Check not found: {code}"
+
+    action = "success"
+    if check.subject or check.subject_fail:
+        action = "ign"
+        subject = email.message_from_string(data).get("subject", "")
+        if check.subject and check.subject in subject:
+            action = "success"
+        elif check.subject_fail and check.subject_fail in subject:
+            action = "fail"
+
+    ua = "Email from %s" % mailfrom
+    check.ping(remote_addr, "email", "", ua, data, action)
+
+    return f"Processed ping for {code}"
+
+
 class Listener(SMTPServer):
     def __init__(self, localaddr, stdout):
         self.stdout = stdout
         super(Listener, self).__init__(localaddr, None, decode_data=False)
 
-    def process_message(
-        self, peer, mailfrom, rcpttos, data, mail_options=None, rcpt_options=None
-    ):
+    def process_message(self, peer, mailfrom, rcpttos, data, **kwargs):
         # get a new db connection in case the old one has timed out:
         connections.close_all()
 
-        to_parts = rcpttos[0].split("@")
-        code = to_parts[0]
-
-        try:
-            data = data.decode()
-        except UnicodeError:
-            data = "[binary data]"
-
-        if not RE_UUID.match(code):
-            self.stdout.write("Not an UUID: %s" % code)
-            return
-
-        try:
-            check = Check.objects.get(code=code)
-        except Check.DoesNotExist:
-            self.stdout.write("Check not found: %s" % code)
-            return
-
-        action = "success"
-        if check.subject:
-            parsed = email.message_from_string(data)
-            received_subject = parsed.get("subject", "")
-            if check.subject not in received_subject:
-                action = "ign"
-
-        ua = "Email from %s" % mailfrom
-        check.ping(peer[0], "email", "", ua, data, action)
-        self.stdout.write("Processed ping for %s" % code)
+        result = _process_message(peer[0], mailfrom, rcpttos[0], data)
+        self.stdout.write(result)
 
 
 class Command(BaseCommand):
@@ -65,6 +69,6 @@ class Command(BaseCommand):
         )
 
     def handle(self, host, port, *args, **options):
-        listener = Listener((host, port), self.stdout)
+        _ = Listener((host, port), self.stdout)
         print("Starting SMTP listener on %s:%d ..." % (host, port))
         asyncore.loop()