+++ hc/front/tests/test_filtering_rules.py
@@ -11,15 +11,21 @@ class FilteringRulesTestCase(BaseTestCase):
         self.redirect_url = "/checks/%s/details/" % self.check.code
 
     def test_it_works(self):
+        payload = {
+            "subject": "SUCCESS",
+            "subject_fail": "ERROR",
+            "methods": "POST",
+            "manual_resume": "1",
+            "filter_by_subject": "yes",
+        }
+
         self.client.login(username="alice@example.org", password="password")
-        r = self.client.post(
-            self.url,
-            data={"subject": "SUCCESS", "methods": "POST", "manual_resume": "1"},
-        )
+        r = self.client.post(self.url, data=payload,)
         self.assertRedirects(r, self.redirect_url)
 
         self.check.refresh_from_db()
         self.assertEqual(self.check.subject, "SUCCESS")
+        self.assertEqual(self.check.subject_fail, "ERROR")
         self.assertEqual(self.check.methods, "POST")
         self.assertTrue(self.check.manual_resume)
 
@@ -27,8 +33,10 @@ class FilteringRulesTestCase(BaseTestCase):
         self.check.method = "POST"
         self.check.save()
 
+        payload = {"subject": "SUCCESS", "methods": "", "filter_by_subject": "yes"}
+
         self.client.login(username="alice@example.org", password="password")
-        r = self.client.post(self.url, data={"subject": "SUCCESS", "methods": ""})
+        r = self.client.post(self.url, data=payload)
         self.assertRedirects(r, self.redirect_url)
 
         self.check.refresh_from_db()
@@ -36,21 +44,30 @@ class FilteringRulesTestCase(BaseTestCase):
 
     def test_it_clears_subject(self):
         self.check.subject = "SUCCESS"
+        self.check.subject_fail = "ERROR"
         self.check.save()
 
+        payload = {
+            "methods": "",
+            "filter_by_subject": "no",
+            "subject": "foo",
+            "subject_fail": "bar",
+        }
+
         self.client.login(username="alice@example.org", password="password")
-        r = self.client.post(self.url, data={"methods": ""})
+        r = self.client.post(self.url, data=payload)
         self.assertRedirects(r, self.redirect_url)
 
         self.check.refresh_from_db()
         self.assertEqual(self.check.subject, "")
+        self.assertEqual(self.check.subject_fail, "")
 
     def test_it_clears_manual_resume_flag(self):
         self.check.manual_resume = True
         self.check.save()
 
         self.client.login(username="alice@example.org", password="password")
-        r = self.client.post(self.url, data={})
+        r = self.client.post(self.url, data={"filter_by_subject": "no"})
         self.assertRedirects(r, self.redirect_url)
 
         self.check.refresh_from_db()