+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Added /api/v1/checks/uuid/flips/ endpoint (#349)
 - In the cron expression dialog, show a human-friendly version of the expression
 - Indicate a started check with a progress spinner under status icon (#338)
+- Added "Docs > Reliability Tips" page
 
 ### Bug Fixes
 - Removing Pager Team integration, project appears to be discontinued